Split single instance checker into platform specific files
