+++ core/CMakeLists.txt
@@ -26,15 +26,18 @@ file(GLOB UTILS_FILES
 if(WIN32)
     list(APPEND UTILS_FILES src/utils/os/DirsWin32.cpp)
     list(APPEND UTILS_FILES src/utils/os/SysWin32.cpp)
+    list(APPEND UTILS_FILES src/utils/os/SingleInstanceCheckerWin32.cpp)
 else()
     list(APPEND UTILS_FILES src/utils/os/DirsPosix.cpp)
     list(APPEND UTILS_FILES src/utils/os/SysPosix.cpp)
     if(APPLE)
         list(APPEND UTILS_FILES src/utils/os/DirsMacOS.cpp)
         list(APPEND UTILS_FILES src/utils/os/SysMacOS.cpp)
+        list(APPEND UTILS_FILES src/utils/os/SingleInstanceCheckerMacOS.cpp)
     else()
         list(APPEND UTILS_FILES src/utils/os/DirsLinux.cpp)
         list(APPEND UTILS_FILES src/utils/os/SysLinux.cpp)
+        list(APPEND UTILS_FILES src/utils/os/SingleInstanceCheckerLinux.cpp)
     endif()
 endif()
