+++ core/include/core/utils/Log.h
@@ -3,9 +3,7 @@
 #include <spdlog/spdlog.h>
 
 #include <filesystem>
-#include <sstream>
 #include <string_view>
-#include <vector>
 
 namespace fs = std::filesystem;
